Rename: VoucherType name 수정, 패키지 변경
