+++ src/main/java/org/prgrms/voucherapplication/repository/voucher/VoucherRepository.java
@@ -1,4 +1,4 @@
-package org.prgrms.voucherapplication.repository;
+package org.prgrms.voucherapplication.repository.voucher;
 
 import org.prgrms.voucherapplication.entity.Voucher;
